Correct data type on SaleInvoiceDetailDto. Add assertions on sale invoice test.

+++ app/DTO/SaleInvoiceDetailDto.php
@@ -4,9 +4,9 @@ namespace App\DTO;
 
 class SaleInvoiceDetailDto {
     public string $item;
-    public int $quantity;
-    public int $price;
-    public int $subtotal;
-    public int $tax;
+    public float $quantity;
+    public float $price;
+    public float $subtotal;
+    public float $tax;
 
 }
